requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
djangorestframework==3.7.3
idna==2.6
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pygame==1.9.3
pytz==2017.3
requests==2.18.4
urllib3==1.22